move installer to appropriate package
